refactor (gravsearch): move AbstractSparqlTransformer to separate file
